Add files via upload
This is an interactive tool for students, instructors, and ecologists to compare, contrast and visualize ecological data in the fastest possible way. The hypothetical mountain data is developed on basis of the author's past experience in Himalayan studies. The histograms, scatterplots, correlation table, correlation plot, regression graph, summary, and data table are dynamically visualized with options to download in different formats. The user has a choice to decide elevation ranges, vegetation types, species richness, and environmental variables for plots. More interestingly, the graphs and plots are provided with customized functions, which can be used to design plots at the user's choice.